requirements.txt
pandas
fastapi 
uvicorn
rpaframework
PyPDF2==2.10.
selenium
html5lib
camelot-py==0.9.0
opencv-python==4.5.5.64